Add files via upload
#CARS
Q. 1) Find all null values in the data. If there is any null value in that column then fill that value with mean of that column.
Q.2) Check what are the types of makes are there in our dataset, and what is the count(occurrence) of each make in the data.
Q.3) Show all the records where origin is Asia or Europe.
Q.4) Remove all the records where weight is above 4000.
Q.5) Increase all the values of MPG_City by 3.